% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxil.R, R/RcppExports.R
\name{Auxiliary functions}
\alias{Auxiliary functions}
\alias{get_ncols}
\alias{get_firstcolumn}
\alias{get_nlines}
\title{Auxiliary functions}
\usage{
get_ncols(file, sep = " ")

get_firstcolumn(fn, class = "integer", ...)

get_nlines(fn)
}
\arguments{
\item{file}{Filename or connection to read from.}

\item{sep}{Character that separate fields.}

\item{fn}{Filename.}

\item{class}{Character vector of classes to use for column(s), see \code{colClasses} in \code{\link[utils]{read.table}}.}

\item{...}{Parameters sent to \code{read.table}.}
}
\value{
\code{get_ncols}: Number of read fields, including ID columns.

\code{get_firstcolumn}: Vector with elements of first column, or \code{data.frame} if \code{class} contains multiple not-\code{"NULL"} elements.

\code{get_nlines}: Number of lines, or \code{0} on error.
}
\description{
Short-hand functions for retriving file information such as number of columns or rows,
or picking the first column (relevant if this contains IDs).

\code{get_ncols} returns number of columns.

\code{get_firstcolumn} is a wrapper for \code{read.table} that allows the user
to easily get the first column(s) by padding \code{colClasses} with \code{"NULL"} character elements to
skip columns.

\code{get_nlines} returns number of lines. Last line, if empty, is ignored.
}
