% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/bind_snps.cpp, R/mutate_chips.R
\name{cbind_snp_files}
\alias{cbind_snp_files}
\alias{cbind_SNP_files}
\title{Column bind files with genotype matrices}
\usage{
cbind_snp_files(fnin, fnout, skiplines = 0L, excludeids = as.integer(c()),
  idwidth = 4L, precision = -1L)

cbind_SNP_files(fnin, fnout, skiplines = 0L, excludeids = as.integer(c()),
  idwidth = 4L, precision = -1L)
}
\arguments{
\item{fnin}{Character vector of filenames to concatenate.}

\item{fnout}{Filename of resulting file.}

\item{skiplines}{Integer, number of lines to skip before outputting.}

\item{excludeids}{Integer vector of first column id's to \emph{exclude} from the output.}

\item{idwidth}{Width of ID column.}

\item{precision}{Number of decimals to print genotypes; 
when \code{-1} (default) genotypes are printed with mixed 
precision. When \code{0}, genotypes are rounded to whole integers.}
}
\value{
Number of lines written.
}
\description{
Does row-wise concatenation of multiple genotype files, 
ignoring the first column in subsequent files.
\strong{NB!} Assumes rows are ordered identically in all files, but will check that
value in first column is identical. If not, corresponding line is skipped in all files.
}
\seealso{
\code{\link{rbind_snp_files}}
}
