% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_phases}
\alias{convert_phases}
\title{Convert phase files to genotype files.}
\usage{
convert_phases(phasefn, genofn, ncol = NULL, nrow = NULL, int = TRUE,
  naval = 9, numeric.format = "5.2")
}
\arguments{
\item{phasefn}{Filename of input file, every two rows are for same animal.}

\item{genofn}{Filename of intended output.}

\item{ncol}{Number of SNPs in file, if \code{NULL} (default), number is estimated from file.}

\item{nrow}{Number of rows to maximally read from \code{phasefn}. If \code{NULL}, no limit is used.}

\item{int}{Logical (default \code{TRUE}), read and write integers.}

\item{naval}{Missing value, default \code{9}.}

\item{numeric.format}{Character describing \code{<width>.<decimals>}. Default \code{'5.2'}. Has no effect when \code{int=TRUE}.}
}
\value{
Number of rows written.
}
\description{
Simply sums every pair of rows together.
}
\details{
A phase file has format similar to SNP files, expect the genotype on each allele are listed on two pairs of rows.
Each individual has therefore two rows, one for each allele, see following example:

\code{Genotype file:}
\tabular{lccccc}{
 1003 \tab 0 \tab 1 \tab 1 \tab 2 \tab 0 \cr
 1004 \tab 1 \tab 1 \tab 0 \tab 1 \tab 0 \cr
}

\code{Phase file:}
\tabular{lccccc}{
 1003 \tab 0 \tab 0 \tab 1 \tab 1 \tab 0 \cr
 1003 \tab 0 \tab 1 \tab 0 \tab 1 \tab 0 \cr
 1004 \tab 0 \tab 1 \tab 0 \tab 0 \tab 0 \cr
 1004 \tab 1 \tab 0 \tab 0 \tab 1 \tab 0 \cr
}

By default assumes genotypes are whole integers (cf. \code{int} argument).
If numeric values are required, e.g. gene dosages from imputation, use \code{int=FALSE}. 
Use \code{numeric.format} to change from default 5 characters width and 2 decimals (\code{"5.2"}). 
NB! Function does not test for validity of this argument, so change with caution!

\strong{Missing values:} Values after summing less than 0 or greater than 2 are assumed as missing and replaced with \code{naval}.
}

