% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_plink}
\alias{convert_plink}
\title{Converts PLINK binary format to flat format}
\usage{
convert_plink(bfile, outfn, na = 9, newID = 0, nlines = NULL,
  fam = NULL, bim = NULL, bed = NULL, countminor = TRUE, maf = 0,
  extract = NULL, exclude = NULL, keep = NULL, remove = NULL,
  method = "simple", fragments = "chr", remerge = TRUE,
  fragmentfns = NULL)
}
\arguments{
\item{bfile}{Filename of PLINK binary files, i.e. without extension.}

\item{outfn}{Filename of new file.}

\item{na}{Missing value.}

\item{newID}{Integer scalar (default \code{0}) for automatically assigning new IDs. See description for more.}

\item{nlines}{Number of lines to process.}

\item{fam}{If binary files have different stems, specify each of them with \code{fam}, \code{bim}, \code{bed}, and set \code{bfile=NULL}.}

\item{bim}{See \code{fam}.}

\item{bed}{See \code{fam}.}

\item{countminor}{Logical: Should the output count minor allele (default), or major allele as \code{plink --recode A}.}

\item{maf}{Numeric, restrict SNPs to SNPs with this frequency.}

\item{extract}{Extract only these SNPs, see Details.}

\item{exclude}{Do not extract these SNPs, see Details.}

\item{keep}{Keep only these samples, see Details.}

\item{remove}{Removes these samples from output, see Details.}

\item{method}{Character, which of following methods to use: \code{simple}, \code{lowmem}, or \code{drymem}. See Details.}

\item{fragments}{\code{"chr"} or integer vector. Only used when \code{method='lowmem'}.}

\item{remerge}{Logical, whether to re-merge fragmented blocks. Only used when \code{method='lowmem'}.}

\item{fragmentfns}{Character vector or function for producing filenames.}
}
\description{
Converts PLINK binary format to flat format
}
\details{
The new integer IDs can be supplied. If not, they will be made for you.
\code{newID} may be an integer vector and will be used as is.
If data.frame with columns \code{famID}, \code{sampID}, and \code{newID}, they will be reordered to match input file.

\strong{\code{method}} \emph{simple} stores entire genotype matrix \emph{in memory}, as PLINK binary files are stored in locus-major mode, 
i.e. first \eqn{m} bits store first locus for all \eqn{n} animals.
Since we are interested in writing out all \eqn{m} loci for each animal, for efficiency we need to read the entire file.
\emph{lowmem} breaks the loci into smaller chunks (e.g. by chromosome), writes each chunk to a file, and merges them back as with \code{\link{cbind_SNPs}}.
\emph{dryrun} does not call the Fortran subroutine, but returns the treated arguments that would have been sent to the subroutine.

For \code{method='lowmem'} use argument \code{fragment} to indicate how the loci are subdivided. 
When \code{fragment='chr'} (case unsensitive), loci are split according to 1st column of .bim file.
If \code{fragment} is a scalar integer, loci are split into this number of blocks.
If an integer vector of same length as \code{ncol}, it directly specifies which block a locus is sent to. \code{max(fragment)} specifies the number of blocks.
}
\section{Filtering loci or samples}{

Filters on loci or samples can be employed in a number of ways; filtering on loci and samples are handled independently.
Inclusion criteria (\code{extract} and \code{keep}) reduces the output to only those loci or samples that pass the criteria.
Exclusion criteria (\code{exclude} and \code{remove}) are applied \emph{after} inclusion criteria, and \emph{reduces} the output further.

\code{extract} and \code{exclude} can be any combination of:
\describe{
 \item{Logical}{Vector of same length as loci in input file.}
 \item{Integer or numeric}{Indicates positional which loci to include or exclude. Numeric vectors are coerced to integer vectors.}
 \item{Character}{Matched against probe IDs, i.e. 2nd column of .bim file.}
}

\code{keep} and \code{remove} are as \code{exctract} and \code{exclude} above, can be a combination of, and can additionally be:
\describe{
 \item{Character}{Matched against both famID or sampID, i.e. 1st and 2nd column of .fam file.}
 \item{List with named elements \code{famID} and/or \code{sampID}}{The named elements are matches against, respectively, the 1st and 2nd column of the .fam file.}
}
}
\references{
\itemize{
 \item PLINK v. 1.07 BED file format: \url{http://pngu.mgh.harvard.edu/~purcell/plink/binary.shtml}
 \item Shaun Purvell and Christopher Chang. \emph{PLINK v. 1.90} \url{https://www.cog-genomics.org/plink2}
 \item Chang CC, Chow CC, Tellier LCAM, Vattikuti S, Purcell SM, Lee JJ (2015) \href{http://www.gigasciencejournal.com/content/4/1/7}{Second-generation PLINK: rising to the challenge of larger and richer datasets.} \emph{GigaScience}, 4.
}
}
\seealso{
\code{convert_plink} is a direct conversion that does not rely on PLINK.
See the alternate \code{\link{convert_plinkA}} which re-formats the output from \code{plink --recode A}.
}

