% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/convert_plinkA.cpp
\name{convert_plinkA}
\alias{convert_plinkA}
\title{Convert PLINK recoded A to SNP file.}
\usage{
convert_plinkA(fnraw, fnout, newID = as.integer(c(1)), nlines = -1L,
  na = "9", idwidth = 4L, precision = 0L)
}
\arguments{
\item{fnraw}{Plink output filename. Most likely \code{plink.raw} if PLINK command line argument \code{--out} is not used.}

\item{fnout}{Filename of new file.}

\item{newID}{Integer scalar (default \code{1}) for first integer ID, or vector of new integer IDs.}

\item{nlines}{Integer, maximum number of pairs of lines to convert. 
When \code{-1} (default), no maximum.}

\item{na}{Character. Numeric value to use for missing values, but passed as a character scalar.}

\item{idwidth}{Width of ID column.}

\item{precision}{Number of decimals to print genotypes; 
when \code{-1} (default) genotypes are printed with mixed 
precision. When \code{0}, genotypes are rounded to whole integers.}
}
\value{
Data.frame with columns \code{famID}, \code{sampID}, and \code{newID}.
}
\description{
Facilitates converting a PLINK binary file to simplified SNP file format.
Requires using PLINK to recode it to the \code{A} format by using command line \code{plink -bfile <file stem> --recode A}.
This function then swiftly strips of first 6 columns (family ID, sample ID, paternal ID, maternal ID, sex, phenotypic record) 
and inserts an integer-based ID column. \code{NA}'s outputted from PLINK are replaced with \code{na} argument.
}
\references{
\itemize{
 \item PLINK. Purcell and Chang. \url{https://www.cog-genomics.org/plink2}
 \item Chang CC, Chow CC, Tellier LCAM, Vattikuti S, Purcell SM, Lee JJ (2015) Second-generation PLINK: rising to the challenge of larger and richer datasets. GigaScience, 4. doi: \href{https://doi.org/10.1186/s13742-015-0047-8}{10.1186/s13742-015-0047-8} \href{http://gigascience.biomedcentral.com/articles/10.1186/s13742-015-0047-8}{link}.
}
}
\seealso{
\code{\link{convert_plink}} is a direct conversion that does not rely on PLINK.
}
