% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R, R/imputation_accuracy.R
\name{phasotogeno}
\alias{imputation_accuracy1}
\alias{imputation_accuracy3}
\alias{phasotogeno}
\alias{phasotogeno_int}
\title{Deprecated functions}
\usage{
phasotogeno(phasefn, genofn, ncol = NULL, nrow = NULL)

phasotogeno_int(phasefn, genofn, ncol = NULL, nrow = NULL)

imputation_accuracy3(truefn, imputefn, nSNPs = NULL, nAnimals = NULL,
  NAval = 9, standardized = TRUE)

imputation_accuracy1(truefn, imputefn, nSNPs = NULL, nAnimals = NULL,
  NAval = 9, standardized = TRUE)
}
\arguments{
\item{phasefn}{Filename of input file, every two rows are for same animal.}

\item{genofn}{Filename of intended output.}

\item{ncol}{Number of columns to read, if \code{NULL} (default), number is estimated from file.}

\item{nrow}{Number of rows to maximally read from \code{phasefn}. If \code{NULL}, no limit is used.}

\item{truefn}{Filename to true matrix. NB! Max. 255 characters!}

\item{imputefn}{Filename to imputed matrix.}

\item{nSNPs}{Integer, number of snps (i.e. number of columns minus 1).
If \code{NULL} (default), detected from \code{truefn}.}

\item{nAnimals}{Integer, number of rows.
If \code{NULL} (default), detected from \code{truefn}.}

\item{NAval}{Integer, value of missing genotype.}

\item{standardized}{Logical, whether to center and scale genotypes by dataset in \code{true}-matrix.
Currently by subtracting column mean and dividing by column standard deviation.}
}
\value{
Number of rows written.
}
\description{
\code{phasotogeno} and \code{phasotogeno_int} has been replaced by \code{\link{convert_phases}}. 
The difference between the two former functions is covered by the \code{int} argument in the new function.

\code{imputation_accuracy1} and \code{imputation_accuracy3} has been replaced by \code{\link{imputation_accuracy}}.
The difference between the two former functions is now covered by the \code{fast}-argument of the latter.
}

