% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_accuracy.R
\name{imputation_accuracy3}
\alias{imputation_accuracy1}
\alias{imputation_accuracy3}
\title{Deprecated function names}
\usage{
imputation_accuracy3(truefn, imputefn, nSNPs = NULL, nAnimals = NULL,
  NAval = 9, standardized = TRUE)

imputation_accuracy1(truefn, imputefn, nSNPs = NULL, nAnimals = NULL,
  NAval = 9, standardized = TRUE)
}
\arguments{
\item{truefn}{Filename to true matrix. NB! Max. 255 characters!}

\item{imputefn}{Filename to imputed matrix.}

\item{nSNPs}{Integer, number of snps (i.e. number of columns minus 1).
If \code{NULL} (default), detected from \code{truefn}.}

\item{nAnimals}{Integer, number of rows.
If \code{NULL} (default), detected from \code{truefn}.}

\item{NAval}{Integer, value of missing genotype.}

\item{standardized}{Logical, whether to center and scale genotypes by dataset in \code{true}-matrix.
Currently by subtracting column mean and dividing by column variance.}
}
\description{
Deprecated function names

Deprecated function names
}

