% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_chips.R
\name{rowconcatenate}
\alias{mergeChips}
\alias{rowconcatenate}
\title{\code{rowconcatenate} has been deprecated.
Use \code{\link{cbind_SNPs}} instead.}
\usage{
rowconcatenate(fns, fnout, nlines = NULL, ncols = NULL, skiplines = 0,
  excludeids = integer(0))

mergeChips(hdid, ldid, hdpos, ldpos, hdfn, ldfn, fnout, outpos = NULL,
  missing = 9)
}
\arguments{
\item{fns}{Character vector of filenames to concatenate. Max 255 characters per filename due to R-Fortran.}

\item{fnout}{Filename of resulting file}

\item{nlines}{Number of lines to read from each file. When \code{NULL} (default), use all lines.}

\item{ncols}{Number of columns to read from each file. When \code{NULL} (default), use all columns.}

\item{skiplines}{Integer, number of lines to skip before outputting. \code{nlines} counts towards total number of lines, i.e. both skipped and outputted.}

\item{excludeids}{Integer vector of first column id's to \emph{exclude} from the output. \code{nlines} also counts towards excluded lines.}

\item{hdid}{IDs of HD genotyped individuals. Corresponds to (subset of) first column of \code{hdfn}.}

\item{ldid}{IDs of LD genotyped individuals. Corresponds to (subset of) first column of \code{ldfn}.}

\item{hdpos}{Integer positions of HD genotyped SNPs. Length must correspond to number of columns in \code{hdfn}, excluding ID column (first column).}

\item{ldpos}{Integer positions of LD genotyped SNPs. Length must correspond to number of columns in \code{hdfn}, excluding ID column (first column).}

\item{hdfn}{Filename of HD genotypes.}

\item{ldfn}{Filename of LD genotypes.}

\item{outpos}{Integer vector of collective SNP positions. Default to sorted, union of \code{hdpos} and \code{ldpos}. Make it anything else and you get?}
}
\description{
\code{rowconcatenate} has been deprecated.
Use \code{\link{cbind_SNPs}} instead.

\code{mergeChips} has been deprecated.
Use \code{\link{rbind_SNPs}} instead.
}

