% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_.R
\name{extract.phased}
\alias{extract.phased}
\title{Extracts}
\usage{
extract.phased(x)
}
\arguments{
\item{x}{Object to extract genotypes from.}
}
\value{
\code{extract.phased} returns matrix with individuals by row, with two rows
  per individual for first and second allele.
}
\description{
Extracts
}
