% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_.R
\name{extract.snps}
\alias{extract.snps}
\alias{extract.snps.haps}
\alias{extract.snps.vcfR}
\alias{extract.snps.oxford}
\title{#' Extracts 
#' 
#' @inheritParams extract.snps
#' @return \code{extract.phased} returns matrix with individuals by row, with two rows
#'   per individual for first and second allele.
extract.phased <- function(x) {
  snps <- t(x$haps)
  rownames(snps) <- rep(rownames(x$samples), each=2)
  snps
}
Extract genotypes from SNP objects, in AlphaImpute format}
\usage{
extract.snps(x, as.integer)

\method{extract.snps}{haps}(x, as.integer = FALSE)

\method{extract.snps}{vcfR}(x, as.integer = FALSE, ...)

\method{extract.snps}{oxford}(x, as.integer = FALSE)
}
\arguments{
\item{x}{Object to extract genotypes from.}

\item{as.integer}{Rounds genotype dosages to whole integers.}

\item{...}{Arguments passed on to \code{vcfR}'s \code{\link[vcfR]{extract.gt}}.}
}
\value{
\code{extract.snps} returns matrix with indivduals by row, with numeric genotypes.
}
\description{
The SNP format lists samples as rows and loci as columns. 
it is simply \code{gt} transposed.
The SNP format is \emph{always} numeric.
}
