% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeit_utils.R
\name{read.haps}
\alias{as.haps}
\alias{is.haps}
\alias{read.haps}
\alias{write.haps}
\title{Reads SHAPEIT haps/sample files}
\usage{
read.haps(haps, sample = NULL)

is.haps(x)

as.haps(x)

write.haps(x, haps, sample)
}
\arguments{
\item{haps}{Filename of .haps file.}

\item{sample}{Filename of .sample file. 
If \code{NULL} (default), uses filestem of \code{haps} appended with \code{.sample}.}

\item{x}{Object to test or convert}
}
\value{
\code{read.haps} returns a list with three entries: 
\describe{
  \item{\code{samples}:}{Data fram with contents of .sample file (three columns).}
  \item{\code{haps}:}{Matrix with unphased genotypes of each allele, with each loci per row and indivduals in tuples of columns.}
  \item{\code{map}:}{Data frame with map of loci, i.e. chromosome, snp ID, chromosomal position, and the two alleles.}
}
}
\description{
Loads SHAPEIT haps/sample into a \code{haps} object.

\code{haps} class requires a list object with three elements, 
\code{haps}, \code{sample}, and \code{map}, with corresponding dimensions.

Write haps object to haps/sample file formats
}
\details{
\code{read.haps} accepts compressed files (.gz and .bz2, see \link[base]{connections}), 
but then both filename arguments must be given.

\code{is.haps} doesn't actually test if \code{x} inherits from class \code{haps},
just that the contents matches.
}
\seealso{
\code{\link{extract.gt}} for vcfR-style extraction of genotypes
}

