% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_chips.R
\name{mask_snp_file}
\alias{mask_snp_file}
\alias{mask_SNPs}
\title{Masks entries of a single SNP chip file.}
\usage{
mask_snp_file(fn, outfn, masking = NULL, snps = NULL, snpsinnew = FALSE,
  dropIDs = NULL, na = 9, int = TRUE, format = NULL)

mask_SNPs(fn, outfn, masking = NULL, snps = NULL, snpsinnew = FALSE,
  dropIDs = NULL, na = 9, int = TRUE, format = NULL)
}
\arguments{
\item{fn}{Input filename.}

\item{outfn}{Output filename.}

\item{masking}{List that specifies how to mask. Positions given here refer to columns in \code{fn}. See details.}

\item{snps}{Vector of column indicies to use in output file. Defaults to all.}

\item{snpsinnew}{Logical, when \code{TRUE}, positions in \code{masking} are mapped to the output file instead of input file.}

\item{dropIDs}{IDs to exclude from output.}

\item{na}{Value to use for masking.}

\item{int}{Logical (default \code{TRUE}), read and write integers.}

\item{format}{Character, Fortran edit descriptors for output. See \link{parseformat}.}
}
\value{
Invisible list of vectors sent to Fortran subtroutine.
}
\description{
This function runs through a \emph{single} file and masks specified columns of specified rows.
It is a simplified version of \code{\link{rbind_snp_files}}.
}
\details{
The \code{masking} argument controls which samples / loci are masked as missing with \code{na}.
It accepts three different objects: 
1) an integer vector of loci to mask, 
2) a list with first element containing integer vector of IDs for samples to mask and second element an integer vector of loci to mask for those samples,
or 
a list of 2).
For 1), the given loci are masked for \emph{all} samples. For 2) it is limited to those IDs given.
By default, the integer vector of loci to mask are mapped to the input file. To map them to output file, set \code{snpsinnew=TRUE}.
}
\examples{

SNPs <- Siccuracy:::make.true(9, 12)
snpfile <- tempfile()
write.snps(SNPs, snpfile)

masking = list(
list(1:3, c(1,3,5,7,9,11)),
list(4:6, c(2,4,6,8,10,12)),
list(7:9, c(1,3,9,10,12))
)

fn <- tempfile()
res <- mask_SNPs(snpfile, fn, masking=masking, na=9)
m <- read.snps(fn)

}
\seealso{
\code{\link{rbind_snp_files}}
}
