% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_chips.R
\name{mask_SNPs}
\alias{mask_SNPs}
\title{Masks entries of a single SNP chip file.}
\usage{
mask_SNPs(fn, outfn, maskIDs, maskSNPs, dropIDs = NULL, dropSNPs = NULL,
  na = 9, ncol = NULL, nlines = NULL, int = TRUE, format = NULL)
}
\arguments{
\item{fn}{Input filename.}

\item{outfn}{Output filename.}

\item{maskIDs}{IDs of rows to mask.}

\item{maskSNPs}{Integer indices of columns in \code{fn} to mask, i.e. before considering \code{dropSNPs}.}

\item{dropIDs}{IDs to exclude from output.}

\item{dropSNPs}{Integer indices of columns in \code{fn} to exclude from output.}

\item{na}{Value to use for masking.}

\item{ncol}{Integer, number of SNP columns in \code{fn}. When \code{NULL} (default), automagically detected with \code{get_ncols(fn)-1}.}

\item{nlines}{Integer, number of lines to process.}

\item{int}{Logical (default \code{TRUE}), read and write integers.}

\item{format}{Character, Fortran edit descriptors for output. See \link{parseformat}.}
}
\description{
This function runs through a single file and masks specified columns of specified rows.
It is a simplified version of \code{\link{rbind_SNPs}}.
}
\details{
If \code{maskIDs} or \code{maskSNPs} are zero-length vectors, \code{dropIDs} and \code{dropSNPs} is still performed.
}

