% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_chips.R
\name{rbind_SNPs}
\alias{rbind_SNPs}
\title{Combines two SNP chip files}
\usage{
rbind_SNPs(hdid, ldid, hdpos, ldpos, hdfn, ldfn, fnout, outpos = NULL,
  na = 9, format = NULL, int = TRUE)
}
\arguments{
\item{hdid}{IDs of HD genotyped individuals. Corresponds to (subset of) first column of \code{hdfn}.}

\item{ldid}{IDs of LD genotyped individuals. Corresponds to (subset of) first column of \code{ldfn}.}

\item{hdpos}{Integer positions of HD genotyped SNPs. Length must correspond to number of columns in \code{hdfn}, excluding ID column (first column).}

\item{ldpos}{Integer positions of LD genotyped SNPs. Length must correspond to number of columns in \code{hdfn}, excluding ID column (first column).}

\item{hdfn}{Filename of HD genotypes.}

\item{ldfn}{Filename of LD genotypes.}

\item{fnout}{Filename to write merged genotypes to.}

\item{outpos}{Integer vector of collective SNP positions. Default to sorted, union of \code{hdpos} and \code{ldpos}. Make it anything else and you get?}

\item{na}{Missing values.}

\item{format}{Character, Fortran edit descriptors for output. See \link{parseformat}.}

\item{int}{Logical (default \code{TRUE}), read and write integers.}
}
\description{
Combines two SNP chip files into one. The two files may be the same, and 
the result will be
}
\details{
We distinguish between \code{HD} and \code{LD} (high and low density) SNP chips,
were HD chips have priority over LD.
}
\seealso{
\code{\link{get_ncols}}
}

