% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxil.R
\name{read.snps}
\alias{read.snps}
\title{Read genotype matrix from file.}
\usage{
read.snps(file, nlines = 0, ncols = NULL, na = NA, what = integer(),
  extractIDs = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{file}{Name of file to read from or connection.}

\item{nlines}{Integer. If positive, the maximum number of line to read.}

\item{ncols}{Integer, default \code{NULL}. Number of columns in input file, including ID column. If \code{NULL} and \code{file} is string, automagically detected.}

\item{na}{If not \code{NA} (default), entries with the value are replaced with \code{NA}.}

\item{what}{Type of internal storage for matrix. Use \code{integer()} or \code{numeric()}, but expect issues if ID column contains alphabetical components.}

\item{extractIDs}{Logical, default \code{TRUE}, trim of first column and use as rownames?}

\item{quiet}{Logical, default \code{TRUE}. If \code{FALSE}, \code{scan} will print a line saying how may items have been read.}

\item{...}{Passed on to \code{\link[base]{scan}}.}
}
\value{
Native \link[base]{matrix}.
}
\description{
\code{read.snps} is wrapper around \code{\link[base]{scan}} that also converts into native matrix format.
}
\details{
Assumes a file format with no header and first column are IDs. If no ID column, use \code{extractIDs = FALSE}.
Usually white-space delimted, but separator can be set with \code{sep} argument as per \code{\link[base]{scan}}.
For file format example see \link{Siccuracy}.
}
\examples{

# Make test data
tmpfile <- tempfile()
write.snps(Siccuracy:::make.true(5, 10), tmpfile)
M <- read.snps(tmpfile)

# Process the genotypes in chunks:
# Setting mode to 'r' is very important to avoid resetting the pointer at file head!
ncols <- get_ncols(tmpfile)
f <- file(tmpfile, 'r') 
while (TRUE) {
  M <- read.snps(f, nlines=500, ncols=ncols)
  if (nrow(M) == 0) break
}
close(f)
}
\seealso{
\code{\link{write.snps}}, \link[base]{typeof}
}

