% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{VCF_Siccuarcy}
\alias{Siccuracy_VCF}
\alias{VCF_Siccuarcy}
\title{VCF objects in Siccuracy}
\description{
To allow fluid interaction with the \emph{Variant Call Format}, 
some functions accepts the objects from the \code[=vcfR]{vcfR} package.
}
\details{
The following functions have been implemented 
with method dispatching for \link[=vcfR]{vcfR} objects:

\describe{
  \item{Imputation_accuracy}{\code{\link[=imputation_accuracy.vcfR]{imputation_accuracy}}}
  \item{Writing SNP data}{\code{\link[=write.snps.vcfR]{write.snps.vcfR}}}
}
}
\author{
Stefan McKinnon Edwards <sme@iysik.com>, August 2017
}

