% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{write.snps.vcfR}
\alias{write.snps.vcfR}
\title{Writes genotypes in VCF class objects to files in \link{AlphaImpute-format}.}
\usage{
\method{write.snps}{vcfR}(x, file, na = "9", newID = 0, element = "GT",
  as.numeric = TRUE, convertNA = TRUE, ...)
}
\arguments{
\item{x}{The matrix to write.}

\item{file}{Filename or connection to write to.}

\item{na}{The string to use for \code{NA} in the data.}

\item{newID}{Integer scalar (default \code{0}) for automatically assigning 
new IDs. Use \code{NULL} to disable. See description for more.}

\item{element}{element to extract from vcf genotype, see \code{\link[vcfR]{extract.gt}}.}

\item{as.numeric}{logical, should the matrix be converted to numerics.}

\item{convertNA}{logical indicating whether to convert "." to \code{NA} 
(see \code{na} for which value to use when writing to text file).}

\item{...}{Arguments forwarded to \code{\link[vcfR]{extract.gt}} and \code{\link{write.snps}}.}
}
\value{
Returns data frame with ID mapping, truncated to those elements outputted.
}
\description{
Writes genotypes in VCF class objects to files in \link{AlphaImpute-format}.
}
\details{
The new integer IDs can be supplied. If not, they will be made for you.
\code{newID} may be an integer vector and will be used as is.
If data.frame with columns \code{sampID} and \code{newID}, they will be reordered to match input file.
}
\seealso{
\code{\link{write.snps}}
}

