% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxil.R
\name{auxfunc}
\alias{auxfunc}
\alias{get_firstcolumn}
\alias{get_ncols}
\alias{get_nlines}
\title{Auxiliary functions}
\usage{
get_ncols(fn)

get_nlines(fn, showWarning = TRUE, doError = FALSE)

get_firstcolumn(fn, class = "integer", ...)
}
\arguments{
\item{fn}{Filename to read from. May be text, gz, or bz2.}

\item{showWarning}{Logical, whether to display a warning if reading file is erroneous.}

\item{doError}{Logical, if error occured while reading file, will it stop}

\item{class}{Character vector of classes to use for column(s), see \code{colClasses} in \code{\link[base]{read.table}}.}

\item{...}{Parameters sent to \code{read.table}.}

\item{nLines}{Optional, number of lines to read. When NULL, \code{get_firstcolumn} uses \code{get_lines} for you.}
}
\value{
\code{get_ncols}: Number of read fields, including ID columns.

\code{get_nlines}: Number of lines.

\code{get_firstcolumn}: Vector with elements of first column, or \code{data.frame} if \code{class} contains multiple not-\code{"NULL"} elements.
}
\description{
Short-hand functions for retriving file information such as number of columns or rows,
or picking the first column (relevant if this contains animal IDs).

\code{get_ncols} works only with numeric values.

\code{get_nlines} works only with numeric values.

\code{get_firstcolumn} is a wrapper for \code{read.table} that allows the user
to easily get the first column(s) by padding \code{colClasses} with \code{"NULL"} character elements to
skip columns.
}

