% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_chips.R
\name{cbind_snp_files}
\alias{cbind_snp_files}
\alias{cbind_SNP_files}
\title{Column bind files with genotype matrices}
\usage{
cbind_snp_files(fns, fnout, ncols = NULL, nlines = NULL, skiplines = 0,
  excludeids = integer(0), format = NULL, int = TRUE)

cbind_SNP_files(...)
}
\arguments{
\item{fns}{Character vector of filenames to concatenate. Max 255 characters per filename due to R-Fortran.}

\item{fnout}{Filename of resulting file}

\item{ncols}{Number of columns to read from each file. When \code{NULL} (default), use all columns.}

\item{nlines}{Number of lines to read from each file. When \code{NULL} (default), use all lines.}

\item{skiplines}{Integer, number of lines to skip before outputting. \code{nlines} counts towards total number of lines, i.e. both skipped and outputted.}

\item{excludeids}{Integer vector of first column id's to \emph{exclude} from the output. \code{nlines} also counts towards excluded lines.}

\item{format}{Character, Fortran edit descriptors for output. See \link{parseformat}.}

\item{int}{Logical (default \code{TRUE}), read and write integers.}

\item{...}{Arguments passed on to \code{cbind_snp_files}
\describe{
  \item{fns}{Character vector of filenames to concatenate. Max 255 characters per filename due to R-Fortran.}
  \item{fnout}{Filename of resulting file}
  \item{nlines}{Number of lines to read from each file. When \code{NULL} (default), use all lines.}
  \item{ncols}{Number of columns to read from each file. When \code{NULL} (default), use all columns.}
  \item{skiplines}{Integer, number of lines to skip before outputting. \code{nlines} counts towards total number of lines, i.e. both skipped and outputted.}
  \item{excludeids}{Integer vector of first column id's to \emph{exclude} from the output. \code{nlines} also counts towards excluded lines.}
  \item{int}{Logical (default \code{TRUE}), read and write integers.}
  \item{format}{Character, Fortran edit descriptors for output. See \link{parseformat}.}
}}
}
\value{
Exit status, invisible.
}
\description{
Does row-wise concatenation of multiple genotype files, 
ignoring the first column in subsequent files.
\strong{NB!} Assumes rows are ordered identically in all files.
}
\seealso{
\code{\link{rbind_snp_files}}
}
