% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/convert_phases.cpp
\name{convert_phases}
\alias{convert_phases}
\title{Convert phase files to genotype files.}
\usage{
convert_phases(fnin, fnout, nlines = -1L, na = 9L, range = as.integer(c(0,
  2)))
}
\arguments{
\item{fnin}{Filename of input file, every two rows are for same animal.}

\item{fnout}{Filename of output file.}

\item{nlines}{Integer, maximum number of pairs of lines to convert. 
When \code{-1} (default), no maximum.}

\item{na}{Value to use for missing values.}

\item{range}{Integer vector of range of allowable values. 
Values in either input or output that are strictly smaller than 
first element or strictly larger than second element are replaced
with \code{na}.}
}
\value{
Number of rows written.
}
\description{
Simply sums every pair of rows together.
}
\details{
A phase file has format similar to SNP files, expect the genotype on each allele are listed on two pairs of rows.
Each individual has therefore two rows, one for each allele, see following example:

\code{Genotype file:}
\tabular{lccccc}{
 1003 \tab 0 \tab 1 \tab 1 \tab 2 \tab 0 \cr
 1004 \tab 1 \tab 1 \tab 0 \tab 1 \tab 0 \cr
}

\code{Phase file:}
\tabular{lccccc}{
 1003 \tab 0 \tab 0 \tab 1 \tab 1 \tab 0 \cr
 1003 \tab 0 \tab 1 \tab 0 \tab 1 \tab 0 \cr
 1004 \tab 0 \tab 1 \tab 0 \tab 0 \tab 0 \cr
 1004 \tab 1 \tab 0 \tab 0 \tab 1 \tab 0 \cr
}

\strong{Missing values:} Values after summing less than 0 or greater than 2 are assumed as missing and replaced with \code{na}.
Value of this range can be changed with argument \code{range}.
}
