% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_phases}
\alias{convert_phases}
\title{Convert phase files to genotype files.}
\usage{
convert_phases(fn, outfn, ncol = NULL, nlines = NULL, na = 9,
  int = TRUE, format = NULL)
}
\arguments{
\item{fn}{Filename of input file, every two rows are for same animal.}

\item{outfn}{Filename of intended output.}

\item{ncol}{Integer, number of SNP columns in files. When \code{NULL}, automagically detected with \code{get_ncols(phasefn)-1}.}

\item{nlines}{Integer, maximum number of pairs of lines to convert.}

\item{na}{Missing value.}

\item{int}{Logical (default \code{TRUE}), read and write integers.}

\item{format}{Character, Fortran edit descriptors for output. See \link{parse.format}.}
}
\value{
Number of rows written.
}
\description{
Simply sums every pair of rows together.
}
\details{
A phase file has format similar to SNP files, expect the genotype on each allele are listed on two pairs of rows.
Each individual has therefore two rows, one for each allele, see following example:

\code{Genotype file:}
\tabular{lccccc}{
 1003 \tab 0 \tab 1 \tab 1 \tab 2 \tab 0 \cr
 1004 \tab 1 \tab 1 \tab 0 \tab 1 \tab 0 \cr
}

\code{Phase file:}
\tabular{lccccc}{
 1003 \tab 0 \tab 0 \tab 1 \tab 1 \tab 0 \cr
 1003 \tab 0 \tab 1 \tab 0 \tab 1 \tab 0 \cr
 1004 \tab 0 \tab 1 \tab 0 \tab 0 \tab 0 \cr
 1004 \tab 1 \tab 0 \tab 0 \tab 1 \tab 0 \cr
}

By default assumes genotypes are whole integers (cf. \code{int} argument).
If numeric values are required, e.g. gene dosages from imputation, use \code{int=FALSE}. 
Use \code{format} to change from default 5 characters width and 2 decimals (\code{"5.2"}). 
NB! Function does not test for validity of this argument, so change with caution!

\strong{Missing values:} Values after summing less than 0 or greater than 2 are assumed as missing and replaced with \code{na}.
}

