% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_plinkA}
\alias{convert_plinkA}
\title{Convert PLINK recoded A to SNP file.}
\usage{
convert_plinkA(rawfn, outfn, newID = 0, ncol = NULL, nlines = NULL,
  naval = 9)
}
\arguments{
\item{rawfn}{Output file from plink. Most likely \code{plink.raw} if PLINK command line argument \code{--output} is not used.}

\item{outfn}{Filename of new file.}

\item{newID}{Integer scalar (default \code{0}) for automatically assigning new IDs. See description for more.}

\item{ncol}{Number of \emph{SNP} columns in \code{rawfn}. Is total number of columns minus 6.}

\item{nlines}{Number of lines to process.}

\item{naval}{Integer scalar to use for \code{NA} values.}
}
\value{
Data.frame with columns \code{famID}, \code{sampID}, and \code{newID}.
}
\description{
Facilitates converting a PLINK binary file to simplified SNP file format.
Requires using PLINK to recode it to the \code{A} format by using command line \code{plink -bfile <file stem> --recode A}.
This function then swiftly strips of first 6 columns (family ID, sample ID, paternal ID, maternal ID, sex, phenotypic record) 
and inserts an integer-based ID column. \code{NA}'s outputted from PLINK are replaced with \code{naval} argument.
}
\details{
The new integer IDs can be supplied. If not, they will be made for you.
\code{newID} may be an integer vector and will be used as is.
If data.frame with columns \code{famID}, \code{sampID}, and \code{newID}, they will be reordered to match input file.
}
\references{
\itemize{
 \item PLINK. Purcell and Chang. \url{https://www.cog-genomics.org/plink2}
 \item \href{http://www.gigasciencejournal.com/content/4/1/7}{Chang CC, Chow CC, Tellier LCAM, Vattikuti S, Purcell SM, Lee JJ (2015) Second-generation PLINK: rising to the challenge of larger and richer datasets. GigaScience, 4.}
}
}

