% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R, R/imputation_accuracy.R, R/mutate_chips.R
\name{phasotogeno}
\alias{imputation_accuracy1}
\alias{imputation_accuracy3}
\alias{mergeChips}
\alias{phasotogeno}
\alias{phasotogeno_int}
\alias{rowconcatenate}
\title{Deprecated functions}
\usage{
phasotogeno(phasefn, genofn, ncol = NULL, nrow = NULL)

phasotogeno_int(phasefn, genofn, ncol = NULL, nrow = NULL)

imputation_accuracy3(truefn, imputefn, nSNPs = NULL, nAnimals = NULL,
  NAval = 9, standardized = TRUE)

imputation_accuracy1(truefn, imputefn, nSNPs = NULL, nAnimals = NULL,
  NAval = 9, standardized = TRUE)

rowconcatenate(fns, fnout, nlines = NULL, ncols = NULL, skiplines = 0,
  excludeids = integer(0))

mergeChips(hdid, ldid, hdpos, ldpos, hdfn, ldfn, fnout, outpos = NULL,
  missing = 9)
}
\arguments{
\item{phasefn}{Filename of input file, every two rows are for same animal.}

\item{genofn}{Filename of intended output.}

\item{ncol}{Number of columns to read, if \code{NULL} (default), number is estimated from file.}

\item{nrow}{Number of rows to maximally read from \code{phasefn}. If \code{NULL}, no limit is used.}

\item{truefn}{Filename of \emph{true} matrix.}

\item{imputefn}{Filename of \emph{imputed} matrix.}

\item{nSNPs}{Deprecated, use \code{ncol}.}

\item{nAnimals}{Deprecated, use \code{nlines}.}

\item{NAval}{Deprecated, use \code{na}.}

\item{standardized}{Logical, whether to center and scale genotypes by dataset in \code{true}-matrix.
Currently by subtracting column mean and dividing by column standard deviation.}

\item{fns}{Character vector of filenames to concatenate. Max 255 characters per filename due to R-Fortran.}

\item{fnout}{Filename of resulting file}

\item{nlines}{Integer, number of lines in \code{truefn}. When \code{NULL}, automagically detected with \code{gen_nlines(truefn)}.}

\item{ncols}{Number of columns to read from each file. When \code{NULL} (default), use all columns.}

\item{skiplines}{Integer, number of lines to skip before outputting. \code{nlines} counts towards total number of lines, i.e. both skipped and outputted.}

\item{excludeids}{Integer vector of first column id's to \emph{exclude} from the output. \code{nlines} also counts towards excluded lines.}

\item{hdid}{IDs of HD genotyped individuals. Corresponds to (subset of) first column of \code{hdfn}.}

\item{ldid}{IDs of LD genotyped individuals. Corresponds to (subset of) first column of \code{ldfn}.}

\item{hdpos}{Integer vector of where columns in \code{hdfn} are positioned in output file. Coerced to same length as columns in \code{hdfn}.}

\item{ldpos}{Integer vector of where columns in \code{ldfn} are positioned in output file. Coerced to same length as columns in \code{ldfn}.}

\item{hdfn}{Filename of HD genotypes.}

\item{ldfn}{Filename of LD genotypes.}

\item{outpos}{Integer vector of collective SNP positions. Default to sorted, union of \code{hdpos} and \code{ldpos}. Make it anything else and you get?}

\item{missing}{Missing value.}
}
\value{
Number of rows written.
}
\description{
\code{phasotogeno} and \code{phasotogeno_int} has been replaced by \code{\link{convert_phases}}. 
The difference between the two former functions is covered by the \code{int} argument in the new function.

\code{imputation_accuracy1} and \code{imputation_accuracy3} has been replaced by \code{\link{imputation_accuracy}}.
The difference between the two former functions is now covered by the \code{adaptive}-argument of the latter.

\code{rowconcatenate} has been deprecated.
Use \code{\link{cbind_SNPs}} instead.

\code{mergeChips} has been deprecated.
Use \code{\link{rbind_SNPs}} instead.
}

