% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_.R
\name{extract.gt}
\alias{extract.gt}
\alias{extract.gt.vcfR}
\alias{extract.gt.haps}
\alias{extract.gt.oxford}
\title{Extract genotypes from SNP objects.}
\usage{
extract.gt(x, as.numeric, ...)

\method{extract.gt}{vcfR}(x, as.numeric, ...)

\method{extract.gt}{haps}(x, as.numeric = TRUE, ...)

\method{extract.gt}{oxford}(x, as.numeric = TRUE, ..., as.integer = FALSE)
}
\arguments{
\item{x}{Object to extract genotypes from.}

\item{as.numeric}{Logical, when \code{TRUE} (default), return numeric matrix
of SNP counts. When \code{FALSE}, character vector with both alleles 
separated by '/'.}

\item{...}{Arguments passed on to \code{vcfR}'s \code{\link[vcfR]{extract.gt}}.}

\item{as.integer}{Rounds genotype dosages to whole integers.}
}
\value{
\code{extract.gt} returns a matrix with loci per row and samples per column.
}
\description{
gt genotypes are stored with loci per row and samples per column.
For the transposed, see \code{\link{extract.snps}}.
}
\seealso{
\code{\link{extract.snps}}
}
