% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeit_utils.R
\name{extract.gt}
\alias{extract.gt}
\alias{extract.gt.haps}
\alias{extract.gt.vcfR}
\alias{extract.phased}
\alias{extract.snps}
\title{Extract genotypes / phases from SNP objects.}
\usage{
extract.gt(x, ...)

\method{extract.gt}{vcfR}(x, ...)

\method{extract.gt}{haps}(x, is.numeric = TRUE, ...)

extract.phased(x)

extract.snps(x)
}
\arguments{
\item{x}{Object to extract genotypes or phases from.}

\item{is.numeric}{Logical, when \code{TRUE} (default), return numeric matrix
of SNP counts. When \code{FALSE}, character vector with both alleles 
separated by '/'.}
}
\value{
Matrix with loci per row and samples per column.

\code{extract.phased} returns individuals by row, with two rows
  per individual for first and second allele.

\code{extract.snps} returns indivduals by row, with numeric genotypes.
}
\description{
Extract genotypes / phases from SNP objects.
}

