% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeit_utils.R
\name{extract.gt}
\alias{extract.gt}
\alias{extract.gt.haps}
\alias{extract.gt.vcfR}
\alias{extract.phased}
\alias{extract.snps}
\alias{extract.snps.haps}
\alias{extract.snps.vcfR}
\title{Extract genotypes / phases from SNP objects.}
\usage{
extract.gt(x, ...)

\method{extract.gt}{vcfR}(x, ...)

\method{extract.gt}{haps}(x, is.numeric = TRUE, ...)

extract.phased(x)

extract.snps(x, ...)

extract.snps.haps(x)

extract.snps.vcfR(x)
}
\arguments{
\item{x}{Object to extract genotypes or phases from.}

\item{is.numeric}{Logical, when \code{TRUE} (default), return numeric matrix
of SNP counts. When \code{FALSE}, character vector with both alleles 
separated by '/'.}
}
\value{
\code{extract.gt} returns a matrix with loci per row and samples per column.

\code{extract.phased} returns matrix with individuals by row, with two rows
  per individual for first and second allele.

\code{extract.snps} returns matrix with indivduals by row, with numeric genotypes.
}
\description{
Extract genotypes / phases from SNP objects.
}

