% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_.R, R/oxford.R
\name{extract.snps}
\alias{extract.snps}
\alias{extract.snps.haps}
\alias{extract.snps.vcfR}
\alias{extract.snps.oxford}
\title{Extract genotypes from SNP objects, in AlphaImpute format}
\usage{
extract.snps(x, ...)

\method{extract.snps}{haps}(x)

\method{extract.snps}{vcfR}(x)

\method{extract.snps}{oxford}(x, as.integer = FALSE)
}
\arguments{
\item{x}{Object to extract genotypes or phases from.}

\item{as.integer}{Rounds genotype dosages to whole integers.}
}
\value{
\code{extract.snps} returns matrix with indivduals by row, with numeric genotypes.
}
\description{
Extract genotypes from SNP objects, in AlphaImpute format
}
