% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_accuracy.R
\name{imputation_accuracy}
\alias{imputation_accuracy}
\title{Imputation accuracy, aka. correlations}
\usage{
imputation_accuracy(truefn, imputefn, ncol = NULL, nlines = NULL, na = 9,
  standardized = TRUE, adaptive = TRUE, center = NULL, scale = NULL,
  p = NULL, excludeIDs = NULL, excludeSNPs = NULL, tol = 0.1)
}
\arguments{
\item{truefn}{Filename of \emph{true} matrix.}

\item{imputefn}{Filename of \emph{imputed} matrix.}

\item{ncol}{Integer, number of SNP columns in files. When \code{NULL}, automagically detected with \code{get_ncols(truefn)-1}.}

\item{nlines}{Integer, number of lines in \code{truefn}. When \code{NULL}, automagically detected with \code{gen_nlines(truefn)}.}

\item{na}{Value of missing genotypes.}

\item{standardized}{Logical, whether to center and scale genotypes by dataset in \code{true}-matrix.
Currently by subtracting column mean and dividing by column standard deviation.}

\item{adaptive}{Use adaptive method (default) that stores \code{truefn} in memory and compares rows by ID in first column.}

\item{center}{Numeric vector of \code{ncol}-length to subtract with for standardization.}

\item{scale}{Numeric vector of \code{ncol}-length to divide by for standardization.}

\item{p}{Shortcut for \code{center} and \code{scale} when using allele frequencies. \code{center=2p} and \code{scale=2p(1-p)}.}

\item{excludeIDs}{Integer vector, exclude these individuals from correlations. \emph{Does not affect calculation of column means and standard deviations.}}

\item{excludeSNPs}{Integer or logical vector, exclude these columns from correlations. \emph{Does not affect calculation of column means and standard deviations.}}
}
\value{
List with following elements:
\describe{
  \item{\code{matcor}}{Matrix-wise correlation between true and imputed matrix.}
  \item{\code{snps}}{Data frame with all snp-wise statistics}
  \item{\code{animals}}{Data frame with all animal-wise statistics}
}
The data frames with statistics consists of columns
\describe{
  \item{\code{rowID}}{Row ID (\code{$animals} only!).}
  \item{\code{means}}{Value subtracted from each column (\code{$snps} only!).}
  \item{\code{sds}}{Value used to scale each column (i.e. standard deviations) (\code{$snps} only!).}
  \item{\code{cors}}{Pearson correlation between true and imputed genotype.}
  \item{\code{correct}}{Number of entries of equal value (within \code{tol})}
  \item{\code{true.na}}{Number of entries in that were missing in \code{truefn} but not \code{imputefn}.}
  \item{\code{imp.na}}{As \code{true.na}, but vice versa.}
  \item{\code{both.na}}{Number of entries that were missing in both files.}
  \item{\code{correct.pct}}{\code{correct} divided by total number of entries bare missing entries in \code{truefn}.}
}
}
\description{
Calculation of column-wise, row-wise, and matrix-wise correlations between
matrix in files \code{truefn} and \code{imputefn}.
Assumes first column in both files is an integer ID column and thus excluded from calculations.
Standardization (subtract mean, divide by standard deviation) is done column-wise based
on means and standard deviations of \code{truefn}.
Correlations are only performed on those rows that are found in \emph{both} files,
based on the first column (ID column).
}
\details{
Genotypes equal to \code{NAval} are considered missing (i.e. \code{NA}) and are not included in the calculations.

This method stores the "true" matrix in memory with a low-precision real type,
and rows in the "imputed" matrix are read and matched by ID.
If there are no extra rows in either matrix and order of IDs is the same,
consider setting \code{adaptive=FALSE}, as this has a memory usage of O(m), compared to O(nm) for the adaptive method, where 'm' is the number of SNPs and 'n' the number of animals.
The non-adaptive method is however, and very surprisingly, slightly slower.
}
\seealso{
\code{\link{write.snps}} for writing SNPs to a file.
}

