% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxford.R
\name{read.oxford}
\alias{is.oxford}
\alias{read.oxford}
\title{Oxford format}
\usage{
read.oxford(file, ...)

is.oxford(x)
}
\arguments{
\item{file}{Filename to read in.}
}
\value{
\code{read.oxford} returns an \code{oxford} object, i.e. a list with two entries: 
\describe{
  \item{\code{map}:}{Data frame with map of loci, i.e. chromosome, snp ID, chromosomal position, and the two alleles.}
  \item{\code{probs}:}{3 dimensional array; rows (d1) are loci, columns (d2) are samples, in three layers comprising
                       gene dosage of homozygote for reference allele, heterozygote, and homozygote for alternative allele.}
}
}
\description{
Oxford format is a flat, single file, with a single loci per row.
The columns are chromosome, SNP ID, (chromosomal) position, 
reference allele and alternative allele. 
Following triplets of columns denote each sample's genotype dosage / probability.
}

