% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_chips.R
\name{rbind_SNPs}
\alias{rbind_SNPs}
\title{Combines two SNP chip files and provides masking.}
\usage{
rbind_SNPs(hdid, ldid, hdpos, ldpos, hdfn, ldfn, fnout, outcol = NULL,
  na = 9, format = NULL, int = TRUE)
}
\arguments{
\item{hdid}{IDs of HD genotyped individuals. Corresponds to (subset of) first column of \code{hdfn}.}

\item{ldid}{IDs of LD genotyped individuals. Corresponds to (subset of) first column of \code{ldfn}.}

\item{hdpos}{Integer vector of where columns in \code{hdfn} are positioned in output file. Coerced to same length as columns in \code{hdfn}.}

\item{ldpos}{Integer vector of where columns in \code{ldfn} are positioned in output file. Coerced to same length as columns in \code{ldfn}.}

\item{hdfn}{Filename of HD genotypes.}

\item{ldfn}{Filename of LD genotypes.}

\item{fnout}{Filename to write merged genotypes to.}

\item{outcol}{Integer, number of columns in output file. When \code{NULL} (default), uses max value of \code{hdpos} and \code{ldpos}.}

\item{na}{Missing values.}

\item{format}{Character, Fortran edit descriptors for output. See \link{parseformat}.}

\item{int}{Logical (default \code{TRUE}), read and write integers.}
}
\value{
Number of lines written, with attribute 'stat' specifying last IO status.
}
\description{
We distinguish between \code{HD} and \code{LD} (high and low density) SNP chips,
were HD chips have priority over LD.
}
\details{
\strong{\code{hdpos} and \code{ldpos}:}
The output file is created with \code{outcol} columns. 
First, rows from \code{hdfn} are added to the file with positions described in \code{hdpos}. The first column is ignored as this is the ID column.
If \code{hdpos=c(2,3,NA,NA,4)}, the first two columns of \code{hdfn} are used for column 2 and 3 in the output, and the last column of \code{hdfn} is used as the 4th.
The 3rd and 4th column in \code{hdfn} are ignored. This allows masking of genotypes.
After \code{hdfn} is processed for rows in \code{hdid}, the same is repeated for \code{ldfn}. There is no restriction for \code{hdfn=ldfn}.
The columns of the output file is prepopulated with \code{na}.
}
\seealso{
\code{\link{get_ncols}}
}

