% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxil.R
\name{read.snps}
\alias{read.snps}
\title{Read genotype matrix from file.}
\usage{
read.snps(file, ncols = NULL, na = NA, what = integer(),
  extractIDs = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{file}{Filename or connection.}

\item{ncols}{Integer, number of columns in \code{file}, including ID column. When \code{NULL} (default), automagically detected with \code{\link{get_ncols}}.}

\item{na}{Missing values; entries with this value are replaced with \code{NA}..}

\item{what}{The \link[base]{typeof} of data to be read, e.g. \code{integer()},  \code{numeric()}, or \code{character()}.}

\item{extractIDs}{Logical, default \code{TRUE}, trim of first column and use as rownames.}

\item{quiet}{logical: if \code{FALSE}, \code{scan()} will print a line, saying how many items have been read.}

\item{...}{Passed on to \code{\link[base]{scan}}.}
}
\value{
Native \link[base]{matrix}.
}
\description{
\code{read.snps} is wrapper around \code{\link[base]{scan}} that also converts into native matrix format.
}
\details{
Assumes a file format with no header and first column are IDs. If no ID column, use \code{extractIDs = FALSE}.
Usually white-space delimted, but separator can be set with \code{sep} argument as per \code{\link[base]{scan}}.
For file format example see \link{Siccuracy}.

If ID columns contains alphabetical elements, use \code{what=character()}. This will however return a character matrix. 
Use e.g. \code{storage.mode(m) <- 'integer'} to convert to integer-based, keeping all other attributes.
}
\examples{

# Make test data
tmpfile <- tempfile()
write.snps(Siccuracy:::make.true(5, 10), tmpfile)
M <- read.snps(tmpfile)

# Process the genotypes in chunks:
# Setting mode to 'r' is very important to avoid resetting the pointer at file head!
ncols <- get_ncols(tmpfile)
f <- file(tmpfile, 'r') 
while (TRUE) {
  M <- read.snps(f, nlines=500, ncols=ncols)
  if (nrow(M) == 0) break
}
close(f)
}
\seealso{
\code{\link{write.snps}}, \link[base]{typeof}, \code{\link{get_ncols}}, \code{\link[base]{scan}},
          and \link{Oxford}, \link{SHAPEIT}, and \link{VCF} formats.
}
