% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Siccuracy.R
\docType{package}
\name{Siccuracy}
\alias{Siccuracy}
\alias{Siccuracy-package}
\title{Stefan's imputation accuracy package}
\description{
Calculation of imputation accuracies and similar allele stats.
}
\author{
Stefan McKinnon Edwards <sme@iysik.com>
}

