% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{VCF_Siccuarcy}
\alias{Siccuracy_VCF}
\alias{VCF_Siccuarcy}
\title{VCF objects in Siccuracy}
\description{
To allow fluid interaction with the \emph{Variant Call Format}, 
some functions accepts the objects from the \link[=vcfR]{vcfR} package.
}
\details{
The following functions have been implemented 
with method dispatching for \link[=vcfR]{vcfR} objects:

\describe{
  \item{Imputation_accuracy}{\code{\link[=imputation_accuracy.vcfR]{imputation_accuracy}}}
  \item{Writing SNP data}{\code{\link[=write.snps.vcfR]{write.snps.vcfR}}}
}
}
\note{
No functions have been implemented for VCF \emph{files}.
The functions above all assume the files have been read in with
\code{\link[vcfR]{read.vcfR}}.

Direct conversion from VCF formatted files to AlphaImpute may be done by
first using PLINK to recode to 'A' format, and then use \code{\link{convert_plinkA}}.
For the first step, PLINK may be called as:

\code{plink --vcf <vcf-filename> --recode A --allow-extra-chr --out temp_file}

and then to convert to AlphaImpute format:

\code{convert_plinkA('temp_file.raw', 'alphaimpute_genotypes.txt')}
}
\author{
Stefan McKinnon Edwards <sme@iysik.com>, August 2017
}

