% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxil.R, R/oxford.R, R/shapeit_utils.R,
%   R/vcf_utils.R
\name{write.snps}
\alias{write.snps}
\alias{write.snps.matrix}
\alias{write.snps.oxford}
\alias{write.snps.haps}
\alias{write.snps.vcfR}
\title{Write genotype matrices to file, AlphaImpute style.}
\usage{
write.snps(x, ...)

\method{write.snps}{matrix}(x, file, row.names = TRUE, na = "9", ...)

\method{write.snps}{oxford}(x, file, row.names = TRUE, na = "9", ...)

\method{write.snps}{haps}(x, file, row.names = TRUE, na = "9",
  phased = FALSE, newID = 0, ...)

\method{write.snps}{vcfR}(x, file, na = "9", newID = 0, element = "GT",
  as.numeric = TRUE, convertNA = TRUE, ...)
}
\arguments{
\item{x}{The matrix to write.}

\item{...}{Passed on to write.table.}

\item{file}{Filename or connection to write to.}

\item{row.names}{If genotype matrix is "raw" and has first column with animals IDs, set this to \code{FALSE}.}

\item{na}{The string to use for \code{NA} in the data.}

\item{phased}{Logical, when \code{FALSE} (default), collapses all loci into single-value
where haplotype information is lost.}

\item{newID}{... New ID, see \code{\link{convert_plinkA}}, but with some caveats here. I don't know.}

\item{element}{element to extract from vcf genotype, see \code{\link[vcfR]{extract.gt}}.}

\item{as.numeric}{logical, should the matrix be converted to numerics.}

\item{convertNA}{logical indicating whether to convert "." to \code{NA} 
(see \code{na} for which value to use when writing to text file).}

\item{row.names}{If genotype matrix is "raw" and has first column with animals IDs, set this to \code{FALSE}.}

\item{row.names}{If genotype matrix is "raw" and has first column with animals IDs, set this to \code{FALSE}.}

\item{newID}{Integer scalar (default \code{0}) for automatically assigning 
new IDs. Use \code{NULL} to disable. See description for more.}

\item{...}{Arguments forwarded to \code{\link[vcfR]{extract.gt}} and \code{\link{write.snps}}.}
}
\value{
Returns data frame with ID mapping, truncated to those elements outputted.
}
\description{
\code{write.snps} is short hand for \code{write.table} with some default options.
For AlphaImpute format see \link{Siccuracy}.

Writes genotypes in VCF class objects to files in \link{AlphaImpute-format}.
}
\details{
The new integer IDs can be supplied. If not, they will be made for you.
\code{newID} may be an integer vector and will be used as is.
If data.frame with columns \code{sampID} and \code{newID}, they will be reordered to match input file.
}
\seealso{
\code{\link[utils]{write.table}}, \link[base]{connections}.

\code{\link{write.snps}}
}
